+++ src/compiler/emitter/mod.rs
@@ -1,4 +1,6 @@
 mod lines_0_1000;
 pub use lines_0_1000::*;
-mod lines_1000_end;
-pub use lines_1000_end::*;
+mod lines_1000_1500;
+pub use lines_1000_1500::*;
+mod lines_1500_end;
+pub use lines_1500_end::*;
